% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffPlot.R
\name{DiffPlot}
\alias{DiffPlot}
\title{Visualization of differential TAD boundaries}
\usage{
DiffPlot(tad_diff, cont_mat1, cont_mat2, resolution, start_coord,
  end_coord, pre_tad = NULL, point_size = 3, palette = "RdYlBu")
}
\arguments{
\item{tad_diff}{Raw object output by TADCompare. Required}

\item{cont_mat1}{contact matrix in either sparse 3 column,
n x n or n x (n+3) form where the first three columns are coordinates in
BED format. If an x n matrix is used, the column names must correspond to
the start point of the corresponding bin. Should correspond
to the first contact matrix input into TADCompare. Required.}

\item{cont_mat2}{contact matrix in either sparse 3 column,
n x n or n x (n+3) form where the first three columns are coordinates in
BED format. If an x n matrix is used, the column names must correspond to
the start point of the corresponding bin. Should correspond
to the second contact matrix input into TADCompare. Required.}

\item{resolution}{Resolution of the data. Required}

\item{start_coord}{The earliest coordinate of the contact matrix to 
plot. Required}

\item{end_coord}{The final coordinate of the contact matrix to 
plot. Required}

\item{pre_tad}{Parameter allowing for drawing of pre-defined TADs. Input
is a list with two entries corresponding to the TADs for contact matrix 1
and contact matrix 2. The matrices must contain columns named "start" and 
"end" corresponding to the start and end of each TAD. For 
accurate annotations, TADCompare results with pre-defined TADs should be
used. Optional}

\item{point_size}{Parameter used to adjust the size of boundary points on 
heatmap plot. Default is 3.}

\item{palette}{Parameter used to adjust color palette. For list of palettes
see https://rdrr.io/cran/RColorBrewer/man/ColorBrewer.html. Default is 
'RdYlBu'}
}
\value{
A plot containing a visualization of the upper diagonal both 
contact matrices with types of non-/differential boundaries labeled.
The first matrix is shown on top and the second on the bottom. If pre_tad
is provided then the outline of the pre-defined TADs are shown. Individual
TAD score and differential TAD scores are shown below the contact matrix
plots.
}
\description{
Visualization of differential TAD boundaries
}
\details{
Given a TADCompare object and two corresponding contact matrices,
Diff_Plot provides visualization of user-specified regions of the genome
with accompanying differential annotations, TAD scores and differential
TAD scores
}
\examples{
#Read in data
data("rao_chr22_prim")
data("rao_chr22_rep")
#Find differential TAD boundaries
tad_diff <- TADCompare(rao_chr22_prim, rao_chr22_rep,
resolution = 50000)
#Create plot
Diff_Plot(tad_diff,cont_mat1, cont_mat2,resolution=50000, 
start_coord=49750000, end_coord=50700000)
}
