% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM12878.40kb.raw.chr_80_480.R
\docType{data}
\name{GM12878.40kb.raw.chr_80_480}
\alias{GM12878.40kb.raw.chr_80_480}
\title{Chromosome 2 contact matrix from Schmitt et al. 2016.}
\format{A data frame with 1001 rows and 1001 variables:
\describe{
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE87112}
}
\usage{
GM12878.40kb.raw.chr_80_480
}
\description{
A 1001x1001 contact matrix from the GM12878 cell line (40kb Resolution)
}
\keyword{datasets}
