% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR90.40kb.raw.chr2.R
\docType{data}
\name{IMR90.40kb.raw.chr2}
\alias{IMR90.40kb.raw.chr2}
\title{A subset of chomosome 2 contact matrix, IMR90 cell line.}
\format{A data frame with 1001 rows and 1001 variables:
\describe{
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE87112}
}
\usage{
IMR90.40kb.raw.chr2
}
\description{
A 1001x1001 contact matrix from the IMR90 cell line, chr2:8000000-48000000,
40kb Resolution, data from Schmitt et al. 2016.
}
\keyword{datasets}
