% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR90.40kb.raw.chr_80_480.R
\docType{data}
\name{IMR90.40kb.raw.chr_80_480}
\alias{IMR90.40kb.raw.chr_80_480}
\title{Chromosome 2 contact matrix from Schmitt et al. 2016.}
\format{A data frame with 1001 rows and 1001 variables:
\describe{
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE87112}
}
\usage{
IMR90.40kb.raw.chr_80_480
}
\description{
A 1001x1001 contact matrix from the IMR90 cell line (40kb Resolution)
}
\keyword{datasets}
