% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rao_chr22_prim.R
\docType{data}
\name{rao_chr22_prim}
\alias{rao_chr22_prim}
\title{Chromosome 22 combined intrachromosomal primary contact matrix from
Rao et al. 2014.}
\format{A data frame with 704 rows and 704 variables:
\describe{
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63525}
}
\usage{
rao_chr22_prim
}
\description{
A 704x704 contact matrix from the GM12878 cell line (50kb Resolution)
}
\keyword{datasets}
