% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rao_chr22_rep.R
\docType{data}
\name{rao_chr22_rep}
\alias{rao_chr22_rep}
\title{Chromosome 22 combined intrachromosomal replicate contact matrix from
Rao et al. 2014.}
\format{A data frame with 704 rows and 704 variables:}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63525}
}
\usage{
rao_chr22_rep
}
\description{
A 704x704 contact matrix from the GM12878 cell line (50kb Resolution)
}
\keyword{datasets}
