% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_mats.R
\docType{data}
\name{time_mats}
\alias{time_mats}
\title{Chromosome 22 time-varying contact matrices from Rao et al. 2014.}
\format{A data frame with 704 rows and 704 variables:
\describe{
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104334}
}
\usage{
time_mats
}
\description{
Four 704x704 contact matrices representing 20,40,60,180 minutes since
auxin treatment and removal from the HCT-116 cell line (50kb Resolution)
}
\keyword{datasets}
