% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCompare.R
\name{TimeCompare}
\alias{TimeCompare}
\title{Time-varying TAD boundary analysis}
\usage{
TimeCompare(cont_mats, resolution, z_thresh = 3, window_size = 15,
  gap_thresh = 0.8, groupings = NULL)
}
\arguments{
\item{cont_mats}{List of contact matrices in either sparse 3 column, n x n or n x (n+3)
form where the first three columns are coordinates in BED format.
If an x n matrix is used, the column names must correspond to the start
point of the corresponding bin. Required.}

\item{resolution}{Resolution of the data. Used to assign TAD boundaries
to genomic regions. If not provided, resolution will be estimated from
column names of the first matrix. Default is "auto"}

\item{z_thresh}{Threshold for boundary score. Higher values result in a
higher threshold for differential TADs. Default is 3.}

\item{gap_thresh}{Required \% of 0s before a region will be considered a gap
and excluded. Default is .8}

\item{groupings}{Vector identifying which group each contact matrix belongs
to. Each group will be combined using consensus boundary scores. Default is
NULL}
}
\value{
A list containing consensus TAD boundaries and overall scores
 \itemize{
 \item Consensus - Data frame containing location of all consensus
 boundaries
 \item All_Regions - Data frame containing consensus scores for all regions
}
}
\description{
Time-varying TAD boundary analysis
}
\details{
Given a list of sparse 3 column, n x n, or n x (n+3) contact
matrices, TimeCompare identifies TADs and classifies time points based on
boundary change. A data frame of time points with at least one boundary
and the corresponding classification is returned.
}
\examples{
#Read in data
data("rao_chr22_rep")
data("rao_chr22_prim")
cont_mats = list(rao_chr22_rep, rao_chr22_prim)
#Find time varying TAD boundaries
diff_list <- TimeCompare(cont_mats, resolution = 50000)
}
