% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTADs.R
\name{ConsensusTADs}
\alias{ConsensusTADs}
\title{Consensus boundary identification}
\usage{
ConsensusTADs(cont_mats, resolution, z_thresh = 3, window_size = 15,
  gap_thresh = 0.8)
}
\arguments{
\item{cont_mats}{List of contact matrices in either sparse 3 column, n x n or n x (n+3)
form where the first three columns are coordinates in BED format.
If an x n matrix is used, the column names must correspond to the start
point of the corresponding bin. Required.}

\item{resolution}{Resolution of the data. Used to assign TAD boundaries
to genomic regions. If not provided, resolution will be estimated from
column names of the first matrix. Default is "auto"}

\item{z_thresh}{Threshold for boundary score. Higher values result in a
higher threshold for differential TADs. Default is 3.}

\item{gap_thresh}{Required \% of 0s before a region will be considered a gap
and excluded. Default is .8}
}
\value{
A list containing consensus TAD boundaries and overall scores
 \itemize{
 \item Consensus - Data frame containing location of all consensus
 boundaries
 \item All_Regions - Data frame containing consensus scores for all regions
}
}
\description{
Consensus boundary identification
}
\details{
Given two sparse 3 column, n x n , or n x (n+3) contact matrices,
TADCompare identifies differential TAD boundaries. Using a novel boundary
score metric, TADCompare simultaneously identifies TAD boundaries and
tests for differential enrichment between datasets. Strength of interactions
are provided using raw boundary scores and p-values.
}
\examples{
#Read in data
data("rao_chr22_rep")
data("rao_chr22_prim")
cont_mats = list(rao_chr22_rep, rao_chr22_prim)
#Find consensus TAD boundaries
diff_list <- TADCompare(cont_mats, resolution = 50000)
}
