% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADCompare.R
\name{TADCompare}
\alias{TADCompare}
\title{Algorithm for differential TAD detection}
\usage{
TADCompare(cont_mat1, cont_mat2, resolution, z_thresh = 2,
  window_size = 2e+06/resolution, gap_thresh = 0.8)
}
\arguments{
\item{cont_mat1}{The first contact matrix to compare in either sparse 3 column, n x n or n x (n+3)
form where the first three columns are coordinates in BED format.
If an x n matrix is used, the column names must correspond to the start
point of the corresponding bin. Required.}

\item{cont_mat2}{The first contact matrix to compare in either sparse 3 column, n x n or n x (n+3)
form where the first three columns are coordinates in BED format.
If an x n matrix is used, the column names must correspond to the start
point of the corresponding bin. Required.}

\item{resolution}{Resolution of the contact matrix. Required}

\item{z_thresh}{Threshold for boundary score. Corresponds to the z-score of
differences between eigenvector gaps}

\item{window_size}{Size of the sliding window. Defaults to 2000000/resolution}

\item{gap_threshold}{Corresponds to the percentage of zeros allowed before
a column/row is removed from the analysis. 1=100\%, .7 = 70\%, etc. Default is 1.}
}
\value{
\item{diff_loci}{Data frame containing differential regions and their
corresponding gap score. Additionally, the matrix in which the TAD boundary
exists is listed.}

\item{sd_diff}{Vector containing all boundary scores}
}
\description{
Algorithm for differential TAD detection
}
\details{
Given two contact matrices, TADCompare returns a dataframe listing
regions which contain differential TADs. It also returns a continuous score
of boundary differences that can be used to compare regions of the genome.
}
\examples{
#Read in data
data("rao_chr20_25_prim")
data("rao_chr20_25_rep")
#Find differential TADs
diff_tads <- TADCompare(rao_chr20_25_prim, rao_chr20_25_rep, resolution = 25000)

}
